+++ changelog.md
@@ -1,5 +1,6 @@
 # v0.1.1
 - improve search a whole lot
+- improve table information for notes and contexts
 
 # v0.1.0
 initial release >:3c